EVA_apps/EdurellVideoAnnotation/requirements.txt
Wikipedia-API
absl-py==0.11.0
appdirs==1.4.4
astunparse==1.6.3
bcrypt~=3.2.0
bert-extractive-summarizer==0.5.1
#bigfloat==0.4.0
blis==0.7.4
boto3==1.16.40
botocore==1.19.40
breadability==0.1.20
cachetools==4.2.0
catalogue==1.0.0
conllu==4.4
cymem==2.0.5
cython==0.29.14
docopt==0.6.2
email_validator
#en-core-web-sm==2.3.1
flask-wtf==0.14.3
flask_bootstrap
flask_login
flask~=1.1.2
flatbuffers==1.12
frozendict==1.2
gast==0.3.3
gdown==3.12.2
gensim==3.8.3
google-auth-oauthlib==0.4.2
google-auth==1.24.0
google-pasta==0.2.0
grpcio==1.32.0
imageio-ffmpeg==0.4.2
importlib-metadata==4.13.0 #needed for spacy that has deprecated features
isodate==0.6.0
itsdangerous~=1.1.0
jmespath==0.10.0
keras-preprocessing==1.1.2
keybert==0.5.0
#lecture-summarizer==0.0.1
lexrank==0.1.0
markdown==3.3.3
markupsafe==2.0.1
mediapipe==0.9.0.1
moviepy==1.0.3
murmurhash==1.0.5
networkx~=2.5.1
nltk~=3.6
numpy~=1.19.2
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
pafy==0.5.5
pandas==1.3.0
#pandas~=1.1.5
path-py==12.5.0
phrasemachine
pillow==9.3.0
pke @ git+https://github.com/boudinfl/pke.git@aa7df17214252b6bab2f1988eba89fdce8050818
pke~=1.8.1
plac==1.1.3
preshed==3.0.5
proglog==0.1.9
protobuf==3.14.0
punctuator==0.9.6
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycocotools==2.0.2
pycountry==20.7.3
pyld==2.0.3
pymongo[srv]
pymongo~=3.11.3
pysrt==1.1.2
pytesseract==0.3.10
pytest
python-igraph
python-rake
pytorch-pretrained-bert==0.6.2
#rdflib-jsonld==0.5.0
rdflib-jsonld==0.6.2
#rdflib==5.0.0
rdflib==6.1.1
requests-oauthlib==1.3.0
requests~=2.25.1
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.43
sentence-transformers
sentencepiece==0.1.91
sklearn==0.0
smart-open==4.0.1
spacy==2.3.5
sparqlwrapper~=1.8.5
speechrecognition==3.8.1
srsly==1.0.5
sumy==0.8.1
tensorboard-plugin-wit==1.7.0
tensorboard==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
theano==1.0.5
thinc==7.4.5
#transformers==4.4.2
unidecode==1.2.0
uritools==3.0.0
urlextract==1.2.0
wasabi==0.8.0
werkzeug~=1.0.1
wikipedia~=1.4.0
wrapt==1.12.1
wtforms==2.3.3
youtube-dl
youtube-transcript-api
